tests/wpt/tests/gamepad/gamepad-supported-by-permissions-policy.html

<!DOCTYPE html>
<title>Test that gamepad is advertised in the feature list</title>
<link rel="help" href="https://w3c.github.io/webappsec-permissions-policy/#dom-permissionspolicy-features">
<link rel="help" href="https://github.com/w3c/encrypted-media/pull/432">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
    assert_in_array('gamepad', document.permissionsPolicy.features());
}, 'document.permissionsPolicy.features should advertise gamepad.');
</script>
